fix: eliminate 6 panic-on-unwrap sites and 3 arithmetic correctness bugs
Replace bare .unwrap() calls with proper error propagation in SFTP and
widget code to prevent panics when handles are used before initialization
or after close. Also fix three subtle arithmetic/logic bugs that could
cause silent data corruption or incorrect behavior.

Panic elimination:
- sftp/file.rs: Extract sender() helper that returns SftpChannelError
  instead of panicking when tx channel is None. Applied to all 8 call
  sites across AsyncRead (poll_read x2) and AsyncWrite (poll_write x2,
  poll_flush x2, poll_close x2), plus set_metadata, metadata, and fsync.
  Uses io::ErrorKind::NotConnected for async trait impls.
- sftp/dir.rs: Replace .unwrap() on tx channel with .ok_or_else() returning
  a descriptive anyhow error for the ReadDir request path.
- termwiz/widgets/mod.rs: Replace .unwrap() on graph.root with .ok_or_else()
  returning a descriptive error when compute_layout is called before any
  widget has been added.

Arithmetic/logic fixes:
- scan_pipeline.rs: Use saturating_add(matched.length) instead of bare +
  to prevent overflow when offset + length exceeds usize::MAX, which would
  wrap to a small value and incorrectly skip the overlap check.
- base91/lib.rs: Cast value to u64 before left-shifting by self.bits to
  prevent truncation when the shift would overflow the original integer
  width, causing corrupted output bytes during flush.
- rangeset/lib.rs: Fix range_subtract boundary computation — use i_start
  and i_end directly instead of r1.end.min(i_start) and r1.end.min(i_end),
  which collapsed the remainder ranges when the intersection extended
  beyond the input range's end.

+++ frankenterm/base91/src/lib.rs
@@ -207,7 +207,8 @@ impl<'a> std::io::Write for Base91Decoder<'a> {
 
     fn flush(&mut self) -> std::io::Result<()> {
         if let Some(value) = self.value.take() {
-            let out: [u8; 1] = [(self.accumulator & 0xff) as u8 | (value << self.bits)];
+            let out: [u8; 1] =
+                [(self.accumulator & 0xff) as u8 | ((value as u64) << self.bits) as u8];
             self.writer.write_all(&out)?;
         }
         self.bits = 0;
